requirements.txt
openai
discord-py-interactions